tests/io_utils: move scsi disk functions into storage
Most of io_utils functions are used only by the storage package.
Let us place them next to their users and in a properly-named file.
